essage

+++ Controllers/PostsController.cs
@@ -1,16 +1,10 @@
-﻿using System;
+﻿using Blog.Models;
+using PagedList;
+using System;
 using System.Collections.Generic;
-using System.Data;
-using System.Data.Entity;
 using System.Linq;
 using System.Net;
-using System.Web;
 using System.Web.Mvc;
-using Blog.Models;
-using Microsoft.AspNet.Identity.EntityFramework;
-using Microsoft.AspNet.Identity;
-using PagedList;
-using PagedList.Mvc;
 
 namespace Blog.Controllers
 {
@@ -20,17 +14,15 @@ namespace Blog.Controllers
 
         // GET: Blogs
 
-       
         public ActionResult Index(string search, string tag, string month, string year, int Page = 1, int PageSize = 5)
         {
             ViewBag.archive = Post.Archive(db.Blogs.ToList()).OrderByDescending(x => x.Key);
-            ViewBag.comments = db.Comments.Take(4);
+            ViewBag.comments = db.Comments.OrderByDescending(x => x.DateAdded).Take(4);
 
-           if (search != null)
+            if (search != null)
             {
-
-               return View(db.Blogs.Where(x => x.Name.Contains(search) || x.Text.Contains(search)).OrderByDescending(s => s.DateAdded).ToPagedList(Page, PageSize));
-              }
+                return View(db.Blogs.Where(x => x.Name.Contains(search) || x.Text.Contains(search)).OrderByDescending(s => s.DateAdded).ToPagedList(Page, PageSize));
+            }
             else if (tag != null)
             {
                 return View(db.Blogs.Where(x => x.Tags.Select(m => m.Name).Contains(tag)).OrderByDescending(s => s.DateAdded).ToPagedList(Page, PageSize));
@@ -38,10 +30,8 @@ namespace Blog.Controllers
             else if (month != null && year != null)
             {
                 return View(db.Blogs.Where(x => x.DateAdded.Year.ToString().Equals(year) && x.DateAdded.Month.ToString().Equals(month)).OrderByDescending(s => s.DateAdded).ToPagedList(Page, PageSize));
-
             }
-          
-            else 
+            else
             {
                 return View(db.Blogs.OrderByDescending(s => s.DateAdded).ToPagedList(Page, PageSize));
             }
@@ -49,10 +39,8 @@ namespace Blog.Controllers
 
         public PartialViewResult GetPartialIndex(string search, string tag, string month, string year, int Page = 1, int PageSize = 5)
         {
-
             if (search != null)
             {
-
                 return PartialView("_Postlist", db.Blogs.Where(x => x.Name.Contains(search) || x.Text.Contains(search)).OrderByDescending(s => s.DateAdded).ToPagedList(Page, PageSize));
             }
             else if (tag != null)
@@ -62,46 +50,35 @@ namespace Blog.Controllers
             else if (month != null && year != null)
             {
                 return PartialView("_Postlist", db.Blogs.Where(x => x.DateAdded.Year.ToString().Equals(year) && x.DateAdded.Month.ToString().Equals(month)).OrderByDescending(s => s.DateAdded).ToPagedList(Page, PageSize));
-
             }
-
             else
             {
                 return PartialView("_Postlist", db.Blogs.OrderByDescending(s => s.DateAdded).ToPagedList(Page, PageSize));
             }
         }
 
-        public ActionResult Commentform(int postid, int parentid)
-        {
-
-            return View("ReplyCommentForm", new Blog.Models.Comment() { PostId = db.Blogs.Find(postid), ParentId = parentid });
-
-        }
-      
-
         [HttpPost]
         public ActionResult AngularTagInit(int id)
         {
-
-
             return Json(new { Result = db.Tags.Where(x => x.PostId.Id == id).Select(x => new { x.Name, x.Id }).ToList() });
-
         }
 
         [HttpPost]
+        [Authorize]
         public ActionResult AngularTagRemove(int tagid, int postid)
         {
-
             Tag tag = db.Tags.Find(tagid);
-
+            if (tag != null)
+            { 
             db.Tags.Remove(tag);
+            }
             db.SaveChanges();
 
             return Json(new { Result = db.Tags.Where(x => x.PostId.Id == postid).Select(x => new { x.Name, x.Id }).ToList() });
-
         }
 
         [HttpPost]
+        [Authorize]
         public ActionResult AngularTag(String value, int id)
         {
             Post blog = db.Blogs.Find(id);
@@ -109,57 +86,45 @@ namespace Blog.Controllers
             tag.PostId = blog;
             tag.Name = value;
             blog.Tags.Add(tag);
-          
+
             db.SaveChanges();
 
-            return Json(new { Result = db.Tags.Where(x => x.PostId.Id ==id).Select(x => new { x.Name, x.Id }).ToList() });
-
+            return Json(new { Result = db.Tags.Where(x => x.PostId.Id == id).Select(x => new { x.Name, x.Id }).ToList() });
         }
-
 
         [HttpPost]
         public ActionResult TagWidget()
         {
-
-     
             return Json(new { Result = db.Tags.GroupBy(x => x.Name).Select(x => new { Name = x.Key, Count = x.Distinct().Count() }).ToList() });
-
         }
-
-       
 
         public ActionResult Search(string search, int Page = 1, int PageSize = 5)
         {
             ApplicationDbContext db = new ApplicationDbContext();
 
             return RedirectToAction("Index", db.Blogs.Where(x => x.Name == search || x.Text == search).OrderByDescending(s => s.DateAdded).ToPagedList(Page, PageSize));
-
         }
 
         // GET: Blogs/Details/5
         public ActionResult Details(int? id, int Page = 1)
         {
             ViewBag.archive = Post.Archive(db.Blogs.ToList()).OrderByDescending(x => x.Key);
-            ViewBag.comments = db.Comments.Take(4);
-          
+            ViewBag.comments = db.Comments.OrderByDescending(x => x.DateAdded).Take(4);
 
-            
             ViewData["Page"] = Page;
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            
+
             Post blog = db.Blogs.Find(id);
-          
+
             if (blog == null)
             {
                 return HttpNotFound();
             }
-          
-               return View(blog);
-           
-            
+
+            return View(blog);
         }
 
         // GET: Blogs/Create
@@ -167,28 +132,28 @@ namespace Blog.Controllers
         public ActionResult Create()
         {
             ViewBag.archive = Post.Archive(db.Blogs.ToList()).OrderByDescending(x => x.Key);
-            ViewBag.comments = db.Comments.Take(4);
+            ViewBag.comments = db.Comments.OrderByDescending(x => x.DateAdded).Take(4);
 
             return View();
         }
 
         // POST: Blogs/Create
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
+        // To protect from overposting attacks, please enable the specific properties you want to bind to, for
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "Id,Name,Text,DateAdded,LastEdit")]  Post blog)
         {
+            ViewBag.archive = Post.Archive(db.Blogs.ToList()).OrderByDescending(x => x.Key);
+            ViewBag.comments = db.Comments.OrderByDescending(x => x.DateAdded).Take(4);
             if (ModelState.IsValid)
             {
                 ApplicationUser user = new ApplicationUser();
                 blog.Authour = user.ApplicationUserName(db);
-               
-               
-               
+
                 blog.DateAdded = DateTime.Now;
-                blog.LastEdit= DateTime.Now;
-              
+                blog.LastEdit = DateTime.Now;
+
                 blog.Authour.Blogs.Add(blog);
                 db.SaveChanges();
                 return RedirectToAction("Index");
@@ -196,10 +161,13 @@ namespace Blog.Controllers
 
             return View(blog);
         }
-          [Authorize(Roles="Admin")]
+
+        [Authorize(Roles = "Admin")]
         // GET: Blogs/Edit/5
         public ActionResult Edit(int? id)
         {
+            ViewBag.archive = Post.Archive(db.Blogs.ToList()).OrderByDescending(x => x.Key);
+            ViewBag.comments = db.Comments.OrderByDescending(x => x.DateAdded).Take(4);
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -213,33 +181,30 @@ namespace Blog.Controllers
         }
 
         // POST: Blogs/Edit/5
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
+        // To protect from overposting attacks, please enable the specific properties you want to bind to, for
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
-         [Authorize(Roles = "Admin")]
+        [Authorize(Roles = "Admin")]
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "Id,Name,Text,DateAdded,LastEdit")] Post blog)
-          {
+        {
+            ViewBag.archive = Post.Archive(db.Blogs.ToList()).OrderByDescending(x => x.Key);
+            ViewBag.comments = db.Comments.OrderByDescending(x => x.DateAdded).Take(4);
 
-              ViewBag.archive = Post.Archive(db.Blogs.ToList()).OrderByDescending(x => x.Key);
-            ViewBag.comments = db.Comments.Take(4);
-
-             if (ModelState.IsValid)
+            if (ModelState.IsValid)
             {
-                DateTime Dateadded = db.Blogs.Find(blog.Id).DateAdded; 
+                DateTime Dateadded = db.Blogs.Find(blog.Id).DateAdded;
                 blog.LastEdit = DateTime.Now;
                 blog.DateAdded = Dateadded;
-                db.Entry(db.Blogs.Find(blog.Id)).CurrentValues.SetValues(blog); 
-              
-              
+                db.Entry(db.Blogs.Find(blog.Id)).CurrentValues.SetValues(blog);
+
                 db.SaveChanges();
                 return RedirectToAction("Index");
             }
             return View(blog);
         }
 
-
-         [Authorize(Roles = "Admin")]
+        [Authorize(Roles = "Admin")]
         public ActionResult Delete(int id)
         {
             Post blog = db.Blogs.Find(id);
